+++ src/DataValue/ScopeDataInterface.php
@@ -6,8 +6,12 @@ interface ScopeDataInterface
 {
     public function getName(): string;
 
+    public function getPairedUpper(): ?string;
+
     public function getScopeLine(): string;
 
      // used to display note
     public function getBaseline(): string;
+
+    public function getPairedLower(): ?string;
 }