+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.1.0]
+### Added
+- Paired scope
+
+### Fixed
+- Label param regexp
+- Variation of display adaptation outscope
+- Limit page height
+
 ## [1.0.1] - 2022-07-06
 ### Added
 - Adaptation of display with outscope notes